revamp migrations: tabel delivery orders, tabel invoices, tabel tanda terima following ERD

+++ laravel/database/migrations/2026_01_01_000002_create_quotations_tables.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->unsignedInteger('id')->unique(); // belum auto_increment dulu
             $table->string('nomor_quotation', 50)->primary();
             $table->unsignedBigInteger('id_staff');
-            $table->unsignedBigInteger('id_client');
+            $table->unsignedInteger('id_client');
             $table->string('nama_project', 255);
             $table->date('tanggal_pembuatan');
             $table->date('valid_sampai');
